+++ src/schematools/cli.py
@@ -19,8 +19,9 @@ import requests
 import sqlalchemy
 from deepdiff import DeepDiff
 from jsonschema import draft7_format_checker
-from sqlalchemy import create_engine, inspect
+from sqlalchemy import inspect
 from sqlalchemy.exc import SQLAlchemyError
+from sqlalchemy.future import create_engine
 from sqlalchemy.schema import CreateTable
 
 from schematools import (
@@ -48,6 +49,7 @@ from schematools.introspect.db import introspect_db_schema
 from schematools.introspect.geojson import introspect_geojson_files
 from schematools.loaders import FileSystemSchemaLoader, get_schema_loader
 from schematools.maps import create_mapfile
+from schematools.naming import to_snake_case, toCamelCase
 from schematools.permissions.db import (
     apply_schema_and_profile_permissions,
     introspect_permissions,
@@ -147,6 +149,12 @@ def import_() -> None:
 @schema.group("export")
 def export() -> None:
     """Subcommand to export data."""
+    pass
+
+
+@schema.group("tocase")
+def tocase() -> None:
+    """Subcommand to make case-changes."""
     pass
 
 
@@ -709,21 +717,25 @@ def show_tablenames(db_url: str) -> None:
 
 @show.command("datasets")
 @option_schema_url
-def show_datasets(schema_url: str) -> None:
+@click.option("--to-snake-case", "snake_it", is_flag=True)
+def show_datasets(schema_url: str, snake_it: bool) -> None:
     """Retrieve the ids of all the datasets."""
     loader = get_schema_loader(schema_url)
+    modifier = to_snake_case if snake_it else lambda x: x
     for dataset_schema in loader.get_all_datasets().values():
-        click.echo(dataset_schema.id)
+        click.echo(modifier(dataset_schema.id))
 
 
 @show.command("datasettables")
 @option_schema_url
 @argument_dataset_id
-def show_datasettables(schema_url: str, dataset_id: str) -> None:
+@click.option("--to-snake-case", "snake_it", is_flag=True)
+def show_datasettables(schema_url: str, dataset_id: str, snake_it: bool) -> None:
     """Retrieve the ids of the datasettables for the indicated dataset."""
     dataset_schema = _get_dataset_schema(dataset_id, schema_url, prefetch_related=False)
+    modifier = to_snake_case if snake_it else lambda x: x
     for dataset_table in dataset_schema.tables:
-        click.echo(dataset_table.id)
+        click.echo(modifier(dataset_table.id))
 
 
 @show.command("mapfile")
@@ -1121,6 +1133,20 @@ def export_jsonls_for(
         export_jsonls(
             connection, dataset_schema, output, table_ids=table_ids, scopes=[], size=size
         )
+
+
+@tocase.command("camel")
+@click.argument("input_str")
+def convert_to_camel_case(input_str: str) -> str:
+    """Converts INPUT_STR to camel case."""
+    click.echo(toCamelCase(input_str))
+
+
+@tocase.command("snake")
+@click.argument("input_str")
+def convert_to_snake_case(input_str: str) -> str:
+    """Converts INPUT_STR to snake case."""
+    click.echo(to_snake_case(input_str))
 
 
 @kafka.command()