+++ src/schematools/importer/base.py
@@ -6,7 +6,7 @@ from contextlib import closing
 from functools import cached_property
 from itertools import islice
 from pathlib import Path
-from typing import Any, Final, Iterator, TypeVar, cast
+from typing import Any, Final, Iterator, TypeVar
 
 import click
 import requests
@@ -179,7 +179,8 @@ class BaseImporter:
                 continue
 
             self.pk_colname_lookup[table_name] = pk_name
-            pks = {getattr(r, pk_name) for r in self.engine.execute(table.select())}
+            with self.engine.connect() as conn:
+                pks = {getattr(r, pk_name) for r in conn.execute(table.select())}
 
             self.pk_values_lookup[table_name] = pks
 
@@ -193,6 +194,7 @@ class BaseImporter:
         ind_extra_index: bool = True,
         limit_tables_to: set | None = None,
         is_versioned_dataset: bool = False,
+        ind_create_pk_lookup: bool = True,
     ) -> None:
         """Generate the tablemodels, tables and indexes.
 
@@ -273,7 +275,8 @@ class BaseImporter:
 
         if ind_tables:
             self.prepare_tables(self.tables, truncate=truncate)
-            self.create_pk_lookup(self.tables)
+            if ind_create_pk_lookup:
+                self.create_pk_lookup(self.tables)
             self.prepare_views()
 
         if ind_extra_index: